Blanket-ban russian links
The bot will now remove all messages containing Russian links. The internal fake_steam_links list is still needed for a user to be timed out.

+++ cogs/mod.py
@@ -1,7 +1,18 @@
 import asyncio
+import re
 
 import discord
 from discord.ext import commands
+
+from cogs.utils import is_mod
+
+# Regex to match a link with '.ru'
+russian_link = re.compile(
+    r'(?:https?:\/\/)?' +
+    r'(?:www\.)?' +
+    r'[-a-zA-Z0-9@:%._\+~#=]{1,256}\.ru' +
+    r'(?:[-a-zA-Z0-9()@:%_\+.~#?&\/=]*)'
+)
 
 
 class AutoMod(commands.Cog):
@@ -10,14 +21,12 @@ class AutoMod(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
+    @is_mod()
     @commands.command()
     async def fakesteam(self, ctx, *, link):
         """Add a link to the list of fake Steam links."""
         if link in self.bot.settings.fake_steam_links:
             return await ctx.send('Link already registered.')
-
-        # Sleep so that we add the link after the on_message below is called
-        await asyncio.sleep(1)
 
         # This way we hit __setattr__
         self.bot.settings.fake_steam_links += [link]
@@ -26,21 +35,38 @@ class AutoMod(commands.Cog):
     @commands.Cog.listener()
     async def on_message(self, message):
         """Detect whether a message contains risky Steam links."""
-        for link in self.bot.settings.fake_steam_links:
-            if link in message.content:
-                break
-        else:
-            # If there was a link in the message, it would've broke
+        if not isinstance(message.author, discord.Member):  # In DMs
             return
 
-        await message.delete()
+        # Ignore moderators
+        if not message.author.guild_permissions.manage_roles:
+            return
+
         if message.webhook_id:  # This was sent by a webhook
             return
 
-        await message.author.add_roles(
-            discord.Object(self.bot.settings.timed_out_role),
-            reason=f'Sent a fake steam link: "{message.content}"'
+        match = re.search(russian_link, message.content)
+        found = None
+        for link in self.bot.settings.fake_steam_links:
+            if link in message.content:
+                found = link
+
+        if not (match or found):
+            return
+
+        await message.delete()
+        await message.channel.send(
+            'Because of a surge in fake steam links, '
+            'we have employed a blankey ban on `.ru` links.',
+            delete_after=5.0
         )
+
+        # Only time out if it is a known link
+        if found:
+            await message.author.add_roles(
+                discord.Object(self.bot.settings.timed_out_role),
+                reason=f'Sent a fake steam link: "{message.content}"'
+            )
 
 
 def setup(bot):
